requirements.txt
streamlit>=1.32
openai>=1.40
python-dotenv>=1.0
pypdf>=4.0
python-docx>=1.1
openpyxl>=3.1
reportlab>=4.0
